requirements.txt
streamlit
opencv-python-headless
pillow
mediapipe
scikit-learn
numpy